Refactoring the kentik specific parts out of main kkc file
